Added: Styles in /rooms, NavBar and minor fix
