+++ client/src/components/rooms/Rooms.jsx
@@ -1,7 +1,10 @@
 import React, {useEffect, useState} from 'react';
 import {useDispatch, useSelector} from 'react-redux'
 
-import Chat from '../Chat';
+import styles from './styles/Rooms.module.css'
+
+import Chat from './Chat';
+import NavBar from '../NavBar';
 
 import JoinRoomForm from './JoinRoomForm';
 import RoomsList from './RoomsList';
@@ -12,15 +15,23 @@ export default function Rooms() {
   const roomId = useSelector(store => store.roomsReducer.roomId)
 
   return (
-    <div>{console.log(isinRoom, roomId)}
+    <div className={styles.mainDiv}>
+      <NavBar />
       {
         isinRoom
         ?
         <Chat roomId={roomId} />
         :
-        <div>
-          <JoinRoomForm />
-          <RoomsList />
+        <div className={styles.subMainDiv}>
+          <div className={styles.lobby}>
+            <div className={styles.div_Chat_Rooms}>
+              <Chat typeofChat={'chatLobby'}/>
+              <div className={styles.div_CreateRoom_RoomsList}>
+                <JoinRoomForm />
+                <RoomsList />
+              </div>
+            </div>
+          </div>
         </div>
       }
     </div>
