--- client/src/components/styles/Chat.module.css
@@ -1,7 +1,0 @@
-.chat{
-    width: 50%;
-    min-height: 50vh;
-    max-height: 50vh;
-    overflow: scroll;
-    border: 2px solid black
-}
