+++ client/src/components/styles/NavBar.module.css
@@ -1,11 +1,25 @@
-.mainDiv{
+/* .mainDiv{
     display: flex;
     justify-content: space-around;
     background: #292b2c;
     padding: 15px 0px 15px 0px;
+} */
+
+.nav{
+    display: flex;
+    justify-content: space-around;
+    background: #292b2c;
+    
 }
 
 .a{
     text-decoration: none;
     color: #878787;
+    margin: auto;
+    font-weight: 300;
+    font-size: 21px;
+}
+
+.logo2{
+    width: 5%;
 }
